requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
tensorflow
keras
flask
requests
pytest
beautifulsoup4
sqlalchemy
psycopg2
jupyter
notebook
black
flake8
mypy
sphinx
pytest-cov
discord.py
python-dotenv
yfinance
xgboost
lightgbm
catboost
statsmodels
plotly
dash
bokeh
fastapi
uvicorn
pydantic
keras-tuner
